# backend/openapi/openapi.snapshot.json
{
  "components": {
    "schemas": {
      "ApprovePartialPermissionRequestDto": {
        "properties": {},
        "type": "object"
      },
...
      "CreateAreaRequestDto": {
        "properties": {},
        "type": "object"
      },
...
      "CreatePermissionRequestDto": {
        "properties": {},
        "type": "object"
      },
...
      "RejectPermissionRequestDto": {
        "properties": {},
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Document Management System API",
    "title": "DMS API",
    "version": "1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/permission-requests": {
      "get": {
        "operationId": "AdminPermissionRequestsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-permission-requests"
        ]
      }
    },
    "/admin/permission-requests/export.csv": {
      "get": {
        "operationId": "AdminPermissionRequestsController_exportCsv",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-permission-requests"
        ]
      }
    },
    "/admin/permission-requests/{id}": {
      "get": {
        "operationId": "AdminPermissionRequestsController_findOne",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-permission-requests"
        ]
      }
    },
    "/admin/permission-requests/{id}/approve": {
      "post": {
        "operationId": "AdminPermissionRequestsController_approve",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-permission-requests"
        ]
      }
    },
    "/admin/permission-requests/{id}/approve-partial": {
      "post": {
        "operationId": "AdminPermissionRequestsController_approvePartial",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApprovePartialPermissionRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-permission-requests"
        ]
      }
    },
    "/admin/permission-requests/{id}/reject": {
      "post": {
        "operationId": "AdminPermissionRequestsController_reject",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectPermissionRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin-permission-requests"
        ]
      }
    },
...
    "/permission-requests": {
      "post": {
        "operationId": "PermissionRequestsController_create",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePermissionRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "permission-requests"
        ]
      }
    },
    "/permission-requests/areas": {
      "post": {
        "operationId": "PermissionRequestsController_createAreaRequest",
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAreaRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "permission-requests"
        ]
      }
    },
    "/permission-requests/mine": {
      "get": {
        "operationId": "PermissionRequestsController_mine",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "permission-requests"
        ]
      }
    },
    "/search": {
      "get": {
        "operationId": "SearchController_search",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "search"
        ]
      }
    },
    "/search/index-status": {
      "get": {
        "operationId": "SearchController_indexStatus",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "search"
        ]
      }
    },
    "/search/reindex": {
      "post": {
        "operationId": "SearchController_reindex",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "search"
        ]
      }
    },
...
  },
  "servers": [],
...
}